// src/utils/logger.d.ts
import debug from 'debug';
export declare const logger: {
    fs: {
        debug: debug.Debugger;
        info: debug.Debugger;
        warn: debug.Debugger;
        error: debug.Debugger;
    };
    fetch: {
        debug: debug.Debugger;
        info: debug.Debugger;
        warn: debug.Debugger;
        error: debug.Debugger;
    };
    vector: {
        debug: debug.Debugger;
        info: debug.Debugger;
        warn: debug.Debugger;
        error: debug.Debugger;
    };
    db: {
        debug: debug.Debugger;
        info: debug.Debugger;
        warn: debug.Debugger;
        error: debug.Debugger;
    };
};
//# sourceMappingURL=logger.d.ts.map